REST API,Services y Repository de Authorization hecho

+++ allmusic/src/main/java/ar/edu/unnoba/poo2024/allmusic/repositories/SongRepository.java
@@ -3,5 +3,5 @@ package ar.edu.unnoba.poo2024.allmusic.repositories;
 import ar.edu.unnoba.poo2024.allmusic.entities.Song;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SongRepository extends JpaRepository<Song, Integer> {
+public interface SongRepository extends JpaRepository<Song, Long> {
 }
